Silver-Tier/requirements.txt
# AI Employee Vault - Silver Tier Dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
watchdog>=4.0.0
plyer>=2.1.0
schedule>=1.2.0
python-dotenv>=1.0.0
playwright>=1.40.0
linkedin-api>=2.0.0
requests>=2.31.0
